Merged PR 47521: Add @method docblocks for higher level PHPStan compatibility
## What's being changed

Two models.

## Why it's being changed

Our Shopware6 integration wants to run PHPStan at level 8, and we're failing with errors like:
```
Call to an undefined method Dotdigital\V3\Models\Contact\Identifiers::getMobileNumber()
```
However, looking at the [levels](https://phpstan.org/user-guide/rule-levels) this feels like good practice even at level 1. The fact is we are deferring method calls that are not explicitly written in code to the `__call` in the AbstractSingletonModel, and as a result we need to indicate what methods are allowable.

We'll have to add more of these as we go.

## How to review / test this change

- See green builds

Related work items: #223915