Merged PR 49155: Allow setDatafields with an empty array
## What's being changed

In this PR we modified the `setDatafields` in order not to create datafield collection if we pass empty array as an argument.

## Why it's being changed

There are cases that we might need to pass empty array as datafields (depending on the implementation logic). In this case the contact creation will fail with message `contacts:invalidValue - The value given for 'dataFields' is invalid`

## How to review / test this change

- test this code
```
    $subscriber = new \Dotdigital\V3\Models\Contact(
        array(
            'matchIdentifier' => 'email',
            'identifiers' => array(
                'email' => 'testprpe@emailsi.io',
            ),
        )
    );
    $subscriber->setDataFields([]);

    $response = $dotdigital->contacts->create($subscriber);
```
- confirm that a contact has been created

Related work items: #232363