Merged PR 56395: Relax array requirement for nested model properties
## What's being changed

We've made some changes to sub-types of Contact and InsightData models.

- We have relaxed the `array` param type in `Dotdigital\V3\Models\Contact\ChannelProperty` and `Dotdigital\V3\Models\InsightData\Record` where properties have their own model type.
- We have added the `array` param type to `Dotdigital\V3\Models\Contact::setPreferences` and `Dotdigital\V3\Models\InsightData::setRecords`.
- We have added setters to some models where we needed the convenience of setting granular properties.

## Why it's being changed

In some of our connector applications we need the flexibility to conditionally set properties on models and their child models, and we'd run into some cases where our code was breaking the SDK.

Consider this code:
```
$emailChannelProperty = new EmailChannelProperty([
    'emailType' => 'html'
]);
if ($optInType) {
    $emailChannelProperty->setOptInType($optInType);
}
$sdkContact->setChannelProperties([
    'email' => $emailChannelProperty
]);
```

When `setChannelProperties` is called, we call `setProperty` for each key in the array. At this point in time, the channel properties for email are already a hydrated `Dotdigital\V3\Models\Contact\ChannelProperties\EmailChannelProperty::class`, causing the native type check for the param to fail.

The addition of `setOptInType` was made in a separate PR, but before we added this, calling the method did nothing (it reaches `__call` in `AbstractSingletonModel` but returns null.

## How to review / test this change

- Generate a valid Order using the `https://r1-api.dotdigital.com/insightData/v3/collections/Orders/schema` for reference.
- Test passing the `ContactIdentity` via the record constructor.
- Test setting the `ContactIdentity` via the `setContactIdentity()` method.
- Test `$dotdigital->insightData->import($data);`.
- Ensure the Import ID is returned (record passing or failing is irrelevant for this test).

## Notes
I've ignored any models where we would never be setting properties from the consumer side on the model. For example, `Dotdigital\V3\Models\Contact\Import\Summary::class` and `Dotdigital\V3\Models\Contact\Import\Failure::class` - the API doesn't present endpoints where we would need to create such data structures.

In some models I've added new setters, but only where the API would allow us to explicitly set properties. So in `Dotdigital\V3\Models\Contact\Preference::class`, we can set the `id` and `isOptedIn`, but not the `publicName` or `privateName`.

Related work items: #263775