Merged PR 58119: Use PHP DateTime for basic date check
## What's being changed

This PR removes the dependency on the nesbot/carbon package, which we were using for ISO8601 date validation. Instead we have implemented a simpler validation using native DateTime. Fixes #1.

## Why it's being changed

It wasn't really in keeping with the spirit of a lightweight SDK to require a hefty package such as Carbon - and it wasn't adding significant value either. The validation we do on data being sent and received doesn't want to be much more than simple type checking.

For dates, the platform requires ISO8601, and naturally has its own validation that will reject incompatible date strings. Our SDK-side validation stops basic incompatibility but will allow some values through that will be rejected by the platform. This is fine.

Not valid and caught SDK-side:
- ''
- 'string'
- '2015-01 first'
- '2011-10-02T23:25:72Z'

Not valid but only caught platform-side:
- '22\t12.78'
- '2000-01-01T01:00:00+1200'

## How to review / test this change

- Test a `contacts->patchByIdentifier` request that includes a consent record (experiment with the date time values)
- Test different values for 'created' or 'modified' parameters in a get contacts parameter collection

Related work items: #268884